+++ src/scalecube-cluster/LogicalCluster/LogicalClusterInternals.js
@@ -1,0 +1,43 @@
+import {MembershipEvent} from "../api/MembershiptEvent";
+import {Cluster} from "../api/Cluster";
+import type {Type} from "../api/MembershiptEvent";
+
+export const CreateLogicalClusterInternals = (internal) => class LogicalClusterInternals{
+    cluster: any;
+
+    constructor(cluster: Cluster){
+        this.cluster = cluster;
+    }
+
+    message(message: MembershipEvent): void {
+        this.cluster.members$.next(message);
+    }
+
+    send(member: Cluster, type: Type, data: any): void {
+        internal(member).message({
+            type: type,
+            id: member.id(),
+            sender: this.cluster.id(),
+            metadata: data
+        });
+    }
+
+    add(members: Cluster[]): void {
+        const self = this.cluster;
+        const data = this.cluster.metadata();
+        members.forEach(
+            (member) => {
+                if (!this.cluster.myMembers[member.id()]) {
+                    this.cluster.myMembers[member.id()] = member;
+                    internal(member).add([self]);
+                    internal(this.cluster).send(member, 'add', data);
+                }
+            }
+        );
+    }
+
+    remove(member: Cluster): void {
+        delete this.cluster.myMembers[member.id()];
+        this.send(member, 'remove', {});
+    }
+};
