+++ src/scalecube-cluster/LogicalCluster/index.js
@@ -1,0 +1,3 @@
+import { LogicalCluster } from './LogicalCluster';
+
+export { LogicalCluster };
