+++ test/scalecube-cluster/cluster.test.js
@@ -1,4 +1,4 @@
-import { LogicalCluster as Cluster } from 'src/scalecube-cluster/LogicalCluster';
+import { LogicalCluster as Cluster } from 'src/scalecube-cluster/LogicalCluster/LogicalCluster';
 import { from } from 'rxjs/observable/from';
 import 'rxjs/add/operator/zip';
 import 'rxjs/add/operator/do';
@@ -11,7 +11,9 @@ describe('Cluster suite', () => {
     const clusterB = new Cluster();
     const clusterC = new Cluster();
 
-
+    clusterA.metadata('clusterA');
+    clusterB.metadata('clusterB');
+    clusterC.metadata('clusterC');
 
     return {clusterA, clusterB, clusterC};
   };
@@ -22,17 +24,14 @@ describe('Cluster suite', () => {
     clusterA.join(clusterB);
     clusterB.join(clusterC);
 
-    expect(clusterA.members()).toEqual([ clusterA, clusterB, clusterC ]);
-    expect(clusterB.members()).toEqual([ clusterB, clusterA, clusterC ]);
-    expect(clusterC.members()).toEqual([ clusterC, clusterB, clusterA ]);
+    expect(clusterA.members().map(i=>i.metadata())).toEqual([ 'clusterA', 'clusterB', 'clusterC' ]);
+    expect(clusterB.members().map(i=>i.metadata())).toEqual([ 'clusterB', 'clusterA', 'clusterC' ]);
+    expect(clusterC.members().map(i=>i.metadata())).toEqual([ 'clusterC', 'clusterB', 'clusterA' ]);
   });
   it('When clusterA shutdown clusterB and C should not have clusterA and remove message should be send', () => {
     const {clusterA, clusterB, clusterC} = createClusters();
     expect.assertions(8);
 
-    clusterA.metadata('clusterA');
-    clusterB.metadata('clusterB');
-    clusterC.metadata('clusterC');
 
     clusterA.join(clusterB);
     clusterB.join(clusterC);
@@ -49,41 +48,48 @@ describe('Cluster suite', () => {
 
     clusterA.shutdown();
 
-    // expect(clusterA).toBe(undefined);
      expect(clusterB.members().map(i=>i.metadata())).toEqual([ clusterB, clusterC ].map(i=>i.metadata()));
      expect(clusterC.members().map(i=>i.metadata())).toEqual([ clusterC, clusterB ].map(i=>i.metadata()));
-    //expect(clusterC.members()).toEqual([ clusterC, clusterB ]);
   });
-  it('Messages are sent', (done) => {
+  it('When A join B and B join C all add essages are sent', (done) => {
     const {clusterA, clusterB, clusterC} = createClusters();
 
     expect.assertions(18);
 
-    clusterA.metadata('clusterA');
-    clusterB.metadata('clusterB');
-    clusterC.metadata('clusterC');
+    const eventFromCluster = 0;
+    const expectValue = 1;
 
     const checkMeessage = (message$AndExpect$) => {
-        expect(message$AndExpect$[ 0 ].metadata).toEqual(message$AndExpect$[ 1 ].data);
-        expect(message$AndExpect$[ 0 ].sender).toEqual(message$AndExpect$[ 1 ].sender);
-        expect(message$AndExpect$[ 0 ].type).toEqual('add');
+        expect(message$AndExpect$[eventFromCluster].metadata).toEqual(message$AndExpect$[expectValue].data);
+        expect(message$AndExpect$[eventFromCluster].sender).toEqual(message$AndExpect$[expectValue].sender);
+        expect(message$AndExpect$[eventFromCluster].type).toEqual('add');
     };
 
     clusterA
       .listenMembership()
-      .zip(from([ {data: "clusterB", sender: clusterB.id()}, {data: "clusterC", sender: clusterC.id()} ]))
+      // zip with expectations
+      .zip(from([
+          {data: "clusterB", sender: clusterB.id()},
+          {data: "clusterC", sender: clusterC.id()}
+          ]))
       .do(checkMeessage)
       .subscribe();
 
     clusterB
       .listenMembership()
-      .zip(from([ {data: "clusterA", sender: clusterA.id()}, {data: "clusterC", sender: clusterC.id()} ]))
+      .zip(from([
+          {data: "clusterA", sender: clusterA.id()},
+          {data: "clusterC", sender: clusterC.id()}
+          ]))
         .do(checkMeessage)
       .subscribe();
 
     clusterC
       .listenMembership()
-      .zip(from([ {data: "clusterB", sender: clusterB.id()}, {data: "clusterA", sender: clusterA.id()} ]))
+      .zip(from([
+          {data: "clusterB", sender: clusterB.id()},
+          {data: "clusterA", sender: clusterA.id()}
+          ]))
         .do(checkMeessage)
       .elementAt(1)
       .do(()=>done())
